Move location of script
