+++ scripts/README.md
@@ -1,0 +1,15 @@
+# Mapanet converter
+
+Convert mapanet JSON dataset to the JSON format we use in Uitdatabank frontend
+
+## Run the script
+
+```shell
+node convert-mapanet-to-udb.mjs INPUT_FILE_NAME OUTPUT_FILE_NAME
+```
+
+e.g.
+
+```shell
+node convert-mapanet-to-udb.mjs GEO-NL-5-STD.json citiesNL.json
+```
